Example app tweaks:  - Centering of package signatures section  - Refactoring
